Entrega 01 Concluída

+++ LojaDeLivros/src/main/java/com/spring/ifpb/repository/EditoraRepository.java
@@ -6,8 +6,8 @@ import org.springframework.stereotype.Repository;
 import com.spring.ifpb.model.Editora;
 
 @Repository
-public interface EditoraRepository extends JpaRepository<Editora, Integer>{
+public interface EditoraRepository extends JpaRepository<Editora, Long>{
 
-	Editora findById(Long id);
+	Editora findById(long id);
 	
 }
